+++ blueprints/flexberry-serializer-init/files/__root__/serializers/__name__.js
@@ -2,5 +2,8 @@ import { Serializer as <%= className %>Serializer } from '../mixins/regenerated/
 import <%if(parentModelName) {%><%=parentClassName%>Serializer from './<%=parentModelName%>';<%}else{%>__ApplicationSerializer from './application';<%}%>
 
 export default <%if(parentModelName) {%><%=parentClassName%>Serializer<%}else{%>__ApplicationSerializer<%}%>.extend(<%= className %>Serializer, {
-
+  /**
+  * Field name where object identifier is kept.
+  */
+  primaryKey: '__PrimaryKey'
 });