+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Blueprints:
     * Add support reexport for addon.
+    * Add generation not stored properties of models.
 
 ## [0.9.0-beta.7] - 2017-04-14
 ### Added