+++ CHANGELOG.md
@@ -3,12 +3,17 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.9.0-beta.8] - 2017-05-05
 ### Added
+* Add `readonly-cell` helper. It can specify `readonly` property for each cell of list components.
 * Blueprints:
     * Add support reexport for addon.
-    * Add generation not stored properties of models.
-    * Add generation of serializers when inheriting models.
-* Add `readonly-cell` helper. It can specify `readonly` property for each cell of list components.
+    * Add generation non-stored properties of models.
+
+### Changed
+* Blueprints:
+    * Now serializers are generating correctly in case of inheriting of models.
 
 ### Fixed
 * `colsconfig-dialog-content` component: